requirements.txt
asgiref==3.3.4
Django==3.2.3
django-debug-toolbar==3.2.1
djangorestframework==3.12.4
Faker==8.1.4
pdfkit==0.6.1
Pillow==8.2.0
python-dateutil==2.8.1
pytz==2021.1
six==1.16.0
sqlparse==0.4.1
text-unidecode==1.3
